requirements.txt
appnope==0.1.0
attrs==20.1.0
Automat==20.2.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.1
cssselect==1.1.0
decorator==4.4.2
fake-useragent==0.1.11
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipython==7.18.1
ipython-genutils==0.2.0
itemadapter==0.1.0
itemloaders==1.0.2
jedi==0.17.2
jmespath==0.10.0
lxml==4.5.2
parsel==1.6.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
Protego==0.1.16
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
queuelib==1.5.0
requests==2.24.0
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
traitlets==5.0.0
Twisted==20.3.0
urllib3==1.25.10
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.1.0
